Separate test and format command in develop (#157)
Because we generally do not want to check formatting
during test. In other words, most of the developers
just want to test it only with test command.

So I separate test and format command in development
environment. But CI test should run with format
test.

As a result, developers can test only with test command
in development environment.
When developer want to pull request, they must fix
formatting issue.

ISSUE=#155